[update] : Service 인터페이스와 구현체 분리하여 추상화
